Split browser search state modules
